Side Product implementation

+++ Farm Animals/Pig.hpp
@@ -8,7 +8,7 @@ class Pig : public MeatProducingAnimal {
         Pig();                      // creates new pig at default position I, J
         std::string getNoise();     // returns animal's noise (pig = "oink oink")
         int getAllowedLand();       // returns index of allowed land type (grassland = 1)
-        int getHungerRate();        // returns rate of hunger increase (pig = 0.6)
+        float getHungerRate();        // returns rate of hunger increase (pig = 0.6)
 };
 
 #endif
